Move unshared routines out of shared.py
